rename to usePublicClientForRegistry and make hooks non default exports
