+++ packages/not-next/src/hooks/useKeycard.ts
@@ -1,0 +1,5 @@
+import { useLocalStorage } from "@uidotdev/usehooks";
+
+export function useKeycard() {
+  return useLocalStorage<`0x${string}` | null>("keyCard", null);
+}
