+++ packages/not-next/src/hooks/useShopId.ts
@@ -2,17 +2,15 @@ import { useEffect, useState } from "react";
 import { useSearch } from "@tanstack/react-router";
 // import { ShopId } from "@/types";
 
-export default function useShopId() {
-  const [shopId, setShopId] = useState(null);
+export function useShopId() {
+  const [shopId, setShopId] = useState<bigint | null>(null);
   // Get the search params
   const search = useSearch({
     from: "/",
   });
 
   useEffect(() => {
-    //If shopId is provided as a query, set it as shopId, otherwise check for storeId in localStorage.
     const id = search.shopId;
-
     if (id) {
       setShopId(BigInt(id));
     }
